Optional data when it exports a file SL #2795
'keep' and 'share-scope' are optional metadata for the activity metadata, so we
should use them only if they are set before and if not we should use default
values for them. This patch fixes it.

For example, when the user downloads a file from Internet only 'keep' is set to
'0' but not 'share-scope'. A similar situation happens when a file is openen
with write from a pen-drive: neither 'keep' nor 'share-scope' are set for the
activity.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Reviewed-By: Sascha Silbe <silbe@activitycentral.com>