Fix collaboration - SL #4482
The problem was related to a race condition between the load of a file,
and the start of the sharing session, that is the reason this was
random, and appear a lot more in xo-4 than in other models.
We used two callback methods, one connected to canvas map-event and
another connected to activity map-event. This is not needed anymore,
and removing the canvas map event callback simplify the operation,
and solves the collaboration.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>