Port from GConf - speech rate and pitch
GConf is part of GTK+ 2 and will not be available.  Write used GConf to
get the speech rate and pitch set by Sugar Frame.

Remove the rate and pitch, and use defaults only.

Due to an AbiWord bug workaround in
1a915240c278b1a4901840e138b25f4c985748bd, and that the AbiWord major and
minor version have not changed since, speech in Write is not enabled
anyway.